Crate tree_sitter_pascal

This crate provides Pascal language support for the tree-sitter parsing library.

Typically, you will use the LANGUAGE constant to add this language to a tree-sitter Parser, and then use the parser to parse some code:

let code = r#"
"#;
let mut parser = tree_sitter::Parser::new();
let language = tree_sitter_pascal::LANGUAGE;
parser
    .set_language(&language.into())
    .expect("Error loading Pascal parser");
let tree = parser.parse(code, None).unwrap();
assert!(!tree.root_node().has_error());

Constants

LANGUAGE

The tree-sitter LanguageFn for this grammar.

NODE_TYPES

The content of the node-types.json file for this grammar.